accountType

+++ src/Api/login/apiLogic.js
@@ -1,5 +1,6 @@
 import axios from "axios"
 import { CorrectModal, IncorrectModal, InfoModal } from "../../components/molecules/modals/Modals";
+import { Decrypt } from "../../utilities/Hooks";
 const urlApi = process.env.REACT_APP_API_URL;
 export const loginConfir = async (data,setName,navigate)=>
 {
@@ -9,7 +10,8 @@ export const loginConfir = async (data,setName,navigate)=>
         const authenticated = axios.post(urlApi+"/api/auth/local",data) 
         .then((res)=>
         {
-            if(res.data.user.confirmed === true){
+            const typeAcc = Decrypt(res.data.user.accounttype)
+            if(res.data.user.confirmed === true && process.env.REACT_APP_ACCOUNTTYPE === typeAcc){
                 setName(res.data)
                 localStorage.setItem("token",res.data.jwt)
                 localStorage.setItem("username", res.data.user.username)
